fixing ts config in packages/html
