+++ packages/html/src/docs/tsconfig.json
@@ -14,5 +14,5 @@
     "baseUrl": "./"
   },
   "exclude": ["node_modules"],
-  "include": ["./src/**/*.ts", "./utils/*.ts"]
+  "include": ["./**/*.ts", "./utils/*.ts"]
 }
